requirements.txt
google-adk[a2a]>=0.1.0
google-genai
uvicorn
fastapi
pydantic
python-dotenv
